api/requirements.txt
anyio==3.7.1 
exceptiongroup==1.1.3 
fastapi==0.104.1 
starlette==0.27.0
python-dotenv==1.0.0
h11==0.14.0 
uvicorn==0.23.2
backports.zoneinfo==0.2.1
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
cffi==1.16.0 
pyjwt==2.8.0
certifi==2023.11.17 
charset-normalizer==3.3.2 
requests==2.31.0 
urllib3==2.1.0